feat: store received SendGrid email in Service Bus for robustness
The HTTP webhook (used by SendGrid) now simply receives the request and extracts its body+headers. It stores that as a Service Bus message which is then picked up by a `ServiceBusTrigger` function. This gives a lot of robustness because Service Bus messages can be automatically redelivered if they failed and can be moved to a "Dead letter" if they fail more than the "Max delivery count".